add mermaid helpers
